Enforce checks in generated script that will remove files

+++ tests/rule_rm_first/references/clean_cmd.bash
@@ -1,5 +1,13 @@
 #!/bin/bash
 
-# Rule : "dir1" "dir2"
-rm dir1/toto.txt
+# Rule : RM_FIRST "dir1" "dir2"
+if [ ! -L dir2/toto.txt -a -f dir2/toto.txt ]
+then
+    rm dir1/toto.txt
+elif [ -L dir2/toto.txt  ]
+then
+    echo "dir2/toto.txt" is a link
+else
+    echo "dir2/toto.txt" do not exist
+fi
 #EOF
