+++ tests/rule_rm_second/references/clean_cmd.bash
@@ -1,5 +1,13 @@
 #!/bin/bash
 
-# Rule : "dir1" "dir2"
-rm dir2/toto.txt
+# Rule : RM_SECOND "dir1" "dir2"
+if [ ! -L dir1/toto.txt -a -f dir1/toto.txt ]
+then
+    rm dir2/toto.txt
+elif [ -L dir1/toto.txt  ]
+then
+    echo "dir1/toto.txt" is a link
+else
+    echo "dir1/toto.txt" do not exist
+fi
 #EOF
